Rename network package to nodes
