+++ core/src/main/java/bisq/core/btc/BitcoinModule.java
@@ -20,9 +20,9 @@ package bisq.core.btc;
 import bisq.core.app.AppOptionKeys;
 import bisq.core.btc.model.AddressEntryList;
 import bisq.core.btc.model.BalanceModel;
-import bisq.core.btc.network.BtcNodes;
-import bisq.core.btc.network.RegTestHost;
-import bisq.core.btc.network.UserAgent;
+import bisq.core.btc.nodes.BtcNodes;
+import bisq.core.btc.nodes.RegTestHost;
+import bisq.core.btc.nodes.UserAgent;
 import bisq.core.btc.setup.WalletsSetup;
 import bisq.core.btc.wallet.BsqCoinSelector;
 import bisq.core.btc.wallet.BsqWalletService;
