+++ core/src/main/java/bisq/core/btc/nodes/ProxySocketFactory.java
@@ -33,7 +33,7 @@
  * PircBotX. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.core.btc.network;
+package bisq.core.btc.nodes;
 
 import javax.net.SocketFactory;
