+++ core/src/main/java/bisq/core/btc/nodes/RegTestHost.java
@@ -15,30 +15,15 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.core.btc.network;
+package bisq.core.btc.nodes;
 
-import javax.inject.Inject;
-import javax.inject.Named;
+public enum RegTestHost {
 
-public class UserAgent {
-    public static final String NAME_KEY = "useragent.name";
-    public static final String VERSION_KEY = "useragent.version";
+    NONE,
+    LOCALHOST,
+    REG_TEST_SERVER; // 188.226.179.109
 
-    private final String name;
-    private final String version;
+    public static final RegTestHost DEFAULT = LOCALHOST;
+    public static final String SERVER_IP = "188.226.179.109";
 
-    @Inject
-    public UserAgent(@Named(NAME_KEY) String name, @Named(VERSION_KEY) String version) {
-        this.name = name;
-        this.version = version;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getVersion() {
-        return version;
-    }
 }
-
