+++ core/src/main/java/bisq/core/user/Preferences.java
@@ -20,8 +20,8 @@ package bisq.core.user;
 import bisq.core.app.AppOptionKeys;
 import bisq.core.app.BisqEnvironment;
 import bisq.core.btc.BtcOptionKeys;
-import bisq.core.btc.network.BaseCurrencyNetwork;
-import bisq.core.btc.network.BtcNodes;
+import bisq.core.btc.nodes.BaseCurrencyNetwork;
+import bisq.core.btc.nodes.BtcNodes;
 import bisq.core.btc.wallet.Restrictions;
 import bisq.core.locale.Country;
 import bisq.core.locale.CountryUtil;
