+++ core/src/test/java/bisq/core/btc/nodes/BtcNetworkConfigTest.java
@@ -15,7 +15,7 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.core.btc.network;
+package bisq.core.btc.nodes;
 
 import bisq.core.btc.setup.WalletConfig;
