+++ core/src/test/java/bisq/core/btc/nodes/BtcNodeConverterTest.java
@@ -15,10 +15,10 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.core.btc.network;
+package bisq.core.btc.nodes;
 
-import bisq.core.btc.network.BtcNodeConverter.Facade;
-import bisq.core.btc.network.BtcNodes.BtcNode;
+import bisq.core.btc.nodes.BtcNodeConverter.Facade;
+import bisq.core.btc.nodes.BtcNodes.BtcNode;
 
 import bisq.network.DnsLookupException;
